refactor(hash-aggr): route grouping sets to the migrated hash streams
`GROUPING SETS` / `CUBE` / `ROLLUP` still fell back to the legacy
`GroupedHashAggregateStream`, because the migrated partial and single
stream predicates required a plain `GROUP BY`. Grouping sets are always
planned with `InputOrderMode::Linear` since #24422, and the shared
`AggregateHashTable` already expands every grouping set of an input
batch, so they can run on `PartialHashAggregateStream` and
`SingleHashAggregateStream`.

- Move `init_empty_grouping_sets` from the partial table into the shared
  table, so single aggregation also emits the grand-total group on empty
  input.
- Drop the `is_single()` condition from the partial and single stream
  predicates. The final and partial-reduce predicates keep it, because
  state-input stages must receive the expanded keys as a plain group by.
- The spill-mode `check_grouping_sets` snapshot changes because the
  migrated partial stream flushes after every batch under memory
  pressure; the final results are unchanged.

Part of #22710
